// frontend/src/app/models/report/reportEnrollmentListData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface ReportEnrollmentListData {
    courseCode?: string;
    first?: number;
    fullCourseEn?: string;
    fullCourseTh?: string;
    fullNameEn?: string;
    fullNameTh?: string;
    memberCourseId?: number;
    memberFirstnameEn?: string;
    memberFirstnameTh?: string;
    memberId?: number;
    memberLastnameEn?: string;
    memberLastnameTh?: string;
    mode?: string;
    paymentAmount?: number;
    payType?: number;
    payTypeEn?: string;
    payTypeTh?: string;
    prefixnameNameEn?: string;
    prefixnameNameTh?: string;
    publicNameEn?: string;
    publicNameTh?: string;
    receiptDate?: string;
    receiptDateList?: string[];
    receiptNo?: string;
    rowNum?: number;
    size?: number;
    studyStatus?: number;
    studyStatusEn?: string;
    studyStatusTh?: string;
    transactionDatetime?: string;
    transactionDatetimeList?: string[];
}